app/requirements.txt
annotated-types==0.6.0
blis==0.7.11
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
google-api-core==2.11.1
google-auth==2.20.0
google-cloud-core==2.3.3
google-cloud-speech==2.20.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.10
numpy==1.26.2
packaging==23.2
preshed==3.0.9
proto-plus==1.22.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
pytube==15.0.0
regex==2023.8.8
requests==2.31.0
rsa==4.9
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.0
srsly==2.4.8
thinc==8.2.2
tiktoken==0.4.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
urllib3==1.26.16
wasabi==1.1.2
weasel==0.3.4
